+++ src/Equinor.Procosys.Preservation.Query/GetTagFunction/RequirementDto.cs
@@ -1,4 +1,4 @@
-﻿namespace Equinor.Procosys.Preservation.Query.TagFunctionAggregate
+﻿namespace Equinor.Procosys.Preservation.Query.GetTagFunction
 {
     public class RequirementDto
     {
