+++ src/tests/Equinor.Procosys.Preservation.Query.Tests/GetTagFunction/RequirementDtoTests.cs
@@ -1,7 +1,7 @@
-﻿using Equinor.Procosys.Preservation.Query.TagFunctionAggregate;
+﻿using Equinor.Procosys.Preservation.Query.GetTagFunction;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace Equinor.Procosys.Preservation.Query.Tests.TagFunctionAggregate
+namespace Equinor.Procosys.Preservation.Query.Tests.GetTagFunction
 {
     [TestClass]
     public class RequirementDtoTests
